backend/requirements.txt
alembic==1.7.3
APScheduler==3.8.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.4
black==21.9b0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
coverage==6.0.1
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
future==0.18.2
geocoder==1.38.1
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.3
packaging==21.0
parameterized==0.8.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
regex==2021.9.30
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-JSON==0.3.3
zipp==3.6.0
